virtual_mouse_requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.7
pyautogui==0.9.54
Pillow==10.0.1
numpy==1.24.3
